feat: ViewModel 적용
